requirements.txt
yfinance
pandas
numpy
matplotlib
fpdf
schedule
python-dotenv
requests
streamlit
pandas
yfinance
schedule
python-dotenv
requests
